+++ crates/bcr-ebill-transport/Cargo.toml
@@ -13,6 +13,10 @@ thiserror.workspace = true
 nostr-sdk.workspace = true
 log.workspace = true
 tokio.workspace = true
+reqwest = { version = "0.12.12", default-features = false, features = [
+  "rustls-tls",
+  "json",
+] }
 
 [dev-dependencies]
 mockall = "0.13.1"
