+++ crates/bcr-ebill-transport/src/email/email.rs
@@ -1,4 +1,4 @@
-use super::Result;
+use crate::Result;
 use async_trait::async_trait;
 #[cfg(test)]
 use mockall::automock;
