+++ crates/bcr-ebill-transport/src/email/email_sendgrid.rs
@@ -1,4 +1,5 @@
-use super::{NotificationEmailTransportApi, Result, email::EmailMessage};
+use super::{NotificationEmailTransportApi, email::EmailMessage};
+use crate::Result;
 use async_trait::async_trait;
 use serde::Serialize;
 use serde_json::Value;
@@ -55,7 +56,7 @@ struct SendgridMessage {
 }
 
 impl TryFrom<EmailMessage> for SendgridMessage {
-    type Error = super::Error;
+    type Error = crate::Error;
     fn try_from(message: EmailMessage) -> Result<Self> {
         let from = SendgridAddress::new(message.from);
         let to = SendgridAddress::new(message.to);
@@ -113,7 +114,6 @@ impl SendgridContent {
 #[cfg(test)]
 mod tests {
 
-    use super::super::test_utils::get_test_email_message;
     use super::*;
 
     #[test]
@@ -136,5 +136,14 @@ mod tests {
             api_key: "api_key".to_string(),
             url: "https://api.sendgrid.com".to_string(),
         });
+    }
+
+    pub fn get_test_email_message() -> EmailMessage {
+        EmailMessage {
+            from: "sender@example.com".to_string(),
+            to: "recipient@example.com".to_string(),
+            subject: "Hello World".to_string(),
+            body: "This is a test email.".to_string(),
+        }
     }
 }
