+++ crates/bcr-ebill-transport/src/email/mod.rs
@@ -1,0 +1,4 @@
+pub mod email;
+pub mod email_sendgrid;
+
+pub use email::NotificationEmailTransportApi;
